pyproject.toml
[project]
name = "osekit"
version = "0.2.5"
description = "OSEkit"
readme = "README.md"
authors = [
    { name = "OSmOSE team", email = "osmose@ensta-bretagne.fr" }
]
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.3",
    "numpy>=2.2.6",
    "pandas>=2.2.3",
    "pytest>=8.3.5",
    "pyyaml>=6.0.2",
    "scipy>=1.15.3",
    "soundfile>=0.13.1",
    "soxr>=0.5.0.post1",
    "tomlkit>=0.13.2",
    "tqdm>=4.67.1",
]

[project.scripts]
OSEkit = "OSmOSE:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "coverage>=7.8.2",
    "myst-parser>=4.0.1",
    "ruff>=0.11.10",
    "sphinx>=8.2.3",
    "sphinx-copybutton>=0.5.2",
    "sphinx-rtd-theme>=3.0.2",
]

[tool.ruff]
lint.select = ["ALL"]

[tool.ruff.lint.extend-per-file-ignores]
"tests/**/*.py" = [
    # at least this three should be fine in tests:
    "S101", # asserts allowed in tests...
    "ARG", # Unused function args -> fixtures nevertheless are functionally relevant...
    "FBT", # Don't care about booleans as positional arguments in tests, e.g. via @pytest.mark.parametrize()
    "D", # Docstring-related stuff
    "SLF001", # Access to private variables
    "BLE001", # Blind exceptions
    "PLR0913", # Too many arguments in methods
]


[tool.pytest.ini_options]
markers = [
    "unit: Mark test as unitary test. Run only these with '-m unit'.",
    "integ: Mark test as integration test. Run only these with '-m integ'.",
    "reg: Mark test as regression test. Run only these with '-m ref'.",
    "allow_log_write_to_file: Disables the automatic patch_filehandlers fixture that prevents logging handlers to write into files."
]
